move undefined var to Resolver
